// src/utils/apis/Users/type.tsx
export interface usersInformation {
  id: string;
  name: string;
  Email: string;
  password: string;
  status: string;
  level: string;
  Department: string;
  Image: string;
  Adhar: string;
  Pan: string;
  Document: string;
  Address: string;
}

export interface usersputInformation {
  id: string;
  name: string;
  Email: string;
  password: string;
  status: string;
  level: string;
  Department: string;
  Image: string;
  Adhar: string;
  Pan: string;
  Document: string;
  Address: string;
}